renaming module from gdpr to consentManagement
